Convert to a custom resource
Resolve foodcritic warnings where we were managing state ourselves.
Remove the incorrect load current resource usage to save variables off
and do default values.
