+++ .kitchen.yml
@@ -9,15 +9,15 @@ verifier:
   name: inspec
 
 platforms:
-  - name: centos-6.8
-  - name: centos-7.3
-  - name: debian-7.11
-  - name: debian-8.6
-  - name: fedora-25
-  - name: opensuse-leap-42.2
+  - name: centos-6
+  - name: centos-7
+  - name: debian-8
+  - name: debian-9
+  - name: fedora-28
+  - name: opensuse-leap-42
   - name: ubuntu-14.04
   - name: ubuntu-16.04
-  - name: freebsd-10.3
+  - name: ubuntu-18.04
   - name: freebsd-11.0
 
 suites:
