+++ test/cookbooks/test/recipes/default.rb
@@ -17,12 +17,12 @@
 # limitations under the License.
 #
 
-apt_update 'update'
+apt_update
 
 include_recipe 'java'
 include_recipe 'ruby_build'
 
-%w(2.4.2 jruby-9.1.13.0).each do |rubie|
+%w(2.4.4 jruby-9.1.13.0).each do |rubie|
   ruby_build_ruby rubie do
     environment('MAKE_OPTS' => "-j #{node['cpu']['total'].to_i + 1}")
   end
